Adding the paginator components
Paginator.js accepts an array of data objects, and prints them in a simple list (using a map list function passed down to the Paginator as props).  It renders the data list with paginator button functionality including prev/next arrows, numbered tabs, and a dropdown to select items per page.